Added dialog to delete/reset buttons

+++ src/app/index/index.component.ts
@@ -1,6 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { UsersService } from '../users.service';
 import { Router } from '@angular/router';
+import { MatDialog, MatDialogConfig } from "@angular/material";
+import { ConfirmationDialogComponent } from "../confirmation-dialog/confirmation-dialog.component";
 
 @Component({
   selector: 'app-index',
@@ -10,7 +12,7 @@ import { Router } from '@angular/router';
 export class IndexComponent implements OnInit {
   title = "Página principal"
 
-  constructor(public router: Router, public usersService: UsersService) { }
+  constructor(private dialog: MatDialog, public router: Router, public usersService: UsersService) { }
 
   ngOnInit() { }
 
@@ -20,10 +22,38 @@ export class IndexComponent implements OnInit {
   }
 
   resetButton() {
-    this.usersService.resetLocalStorage();
+    const dialogConfig = new MatDialogConfig();
+    dialogConfig.autoFocus = false;
+    dialogConfig.data = {
+      message: "Você deseja resetar a lista de usuários para a configuração padrão?",
+      btn1: "Resetar lista de usuários",
+      btn2: 'Cancelar'
+    }
+    var dialogRef = this.dialog.open(ConfirmationDialogComponent, dialogConfig);
+    dialogRef.afterClosed().subscribe(
+      data => {
+        if (data) {
+          this.usersService.resetLocalStorage();
+        }
+      }
+    )    
   }
 
   deleteButton() {
-    this.usersService.deleteLocalStorage();
+    const dialogConfig = new MatDialogConfig();
+    dialogConfig.autoFocus = false;
+    dialogConfig.data = {
+      message: "Você deseja remover todos os usuários cadastrados?",
+      btn1: "Remover todos os usuários",
+      btn2: 'Cancelar'
+    }
+    var dialogRef = this.dialog.open(ConfirmationDialogComponent, dialogConfig);
+    dialogRef.afterClosed().subscribe(
+      data => {
+        if (data) {
+          this.usersService.deleteLocalStorage();
+        }
+      }
+    )
   }
 }
